filter out module overrides from calendar event overrides
Closes LX-2129

Test Plan:

Create course with a module and an assignment and at least one student
in a section.

Create a module override for the section with the student.

Add a date for the section override for that assignment.

Masquerade as the student and go to their calendar.  Ensure that they
receive the calendar event for that assignment on the correct date.
Ensure that the calendar event API call does not include the course
module override in the list of overrides for that event.

Change-Id: Ib238da931ceaf93eedc58639b498e66270de82d0
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/364190
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Sarah Gerard <sarah.gerard@instructure.com>
QA-Review: Sarah Gerard <sarah.gerard@instructure.com>
Product-Review: Robin Kuss <rkuss@instructure.com>