implement a cache key tag for password complexity
This change adds a cache key tag for password complexity to the
password policy. The key tag is to ensure related keys are stored on
the same node (for distributed Redis e.g. Redis Cluster). This will
help with cache invalidation, commands that require related keys to be
stored together, and reduce the number of keys stored.

Also updated the error handling for cases if/when there's a Redis
connection error or we're attempting to execute a command that
expects the current node to be the same as the node where the key is stored.

fixes FOO-4974
flag = password_complexity

test plan:
- enable the feature flag for password_complexity
- as a user, set your password
- create a TXT file with the password you just set on a single line in
  the file and save it
- go to the account authentication setting page
- click “View Options” under the canvas auth provider
- upload the txt file as a forbidden words list and save your changes
- as the same user you set the password on, try changing your password
  to anything else that is not in the list
- verify that the password change is successful
- try changing your password to the password you set in the txt file
- verify that the password change is unsuccessful

Change-Id: Ica33cfbfeffd93ddc6a3d1e9bc1543dde16e7182
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/364774
Reviewed-by: Adrián Grúber <adrian.gruber@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>