add extensive metrics around login flows
closes AE-1141

In general, keep track of 5 actions:
 * starting a new login flow
 * actual login attempts
 * starting an SLO
 * SLO responses
 * auth provider initiated logouts

For all of the above, keep track of all attempts, number of
successes, and number of failures (with reason). Also tag with domain,
auth_type (which may sometimes be a generic type like "oauth2" if
we don't yet know specifically which provider it is), and auth_provider_id
if we have it.

Change-Id: If54fec633efdd3e831013f1a0bd754b232c14c11
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/364779
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Michael Hulse <michael.hulse@instructure.com>
QA-Review: Cody Cutrer <cody@instructure.com>
Product-Review: Cody Cutrer <cody@instructure.com>