Fix focus when removing filter
test plan: Go to Discover page and set filters.
Navigate to a filter with Tab and press enter to remove it.
The focus should go to the previous filter.
Remove the first filter, if there are more the focus
should go to the right, if that was the last filter
the focus should go to the filters button .

flag=lti_registrations_page
refs LRN-9244

Change-Id: I60b944d796be533c9849d1724646d74d7691ac73
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/364976
Reviewed-by: Panna Kristóf <panna.kristof@instructure.com>
QA-Review: Panna Kristóf <panna.kristof@instructure.com>
Product-Review: Panna Kristóf <panna.kristof@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>