create proper list of overrides when module is only override
Test Plan:

1. Create a course with two sections - A and B, a teacher enrolled in both sections, and one student enrolled in each section.

- Create an assignment in the course assigned only to section A with a due date
- Create a module and add the assignment to the module.
- Add an Assign To for the module assigning the module to both sections.
- Access the calendar as the teacher or student in section B and note that the assignment does not display. In the console in dev tools there is an Uncaught TypeError: Cannot read properties of undefined (reading 'id') error.
- Access the calendar as the student in section A and note that the assignment displays on the calendar.

2. Create course with a module and an assignment and at least one student
in a section.

Create a module override for the section with the student.

Add a date for the section override for that assignment.

Masquerade as the student and go to their calendar.  Ensure that they
receive the calendar event for that assignment on the correct date.
Ensure that the calendar event API call does not include the course
module override in the list of overrides for that event.

Closes LX-2183, LX-2187

Change-Id: If77d72ff3dfd7717c9544ba0bb8c2a1d32a6676d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/366609
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Sarah Gerard <sarah.gerard@instructure.com>
QA-Review: Sarah Gerard <sarah.gerard@instructure.com>
Product-Review: Robin Kuss <rkuss@instructure.com>