prefer active role when retrieving role_by_name
This change fixes a bug where the role_by_name method would return the
first role with the given name, regardless of whether it was active or
not. This change modifies the method to prefer the active role with the
given name, otherwise inactive if that role is part of the relation,
or nil if no role with that name exists.

fixes FOO-5016
closes FOO-5021
flag = none

test plan:
- create a custom course role within root account permissions
  i.e. "Course Admin"
- retire that role by deleting it from the permissions UI
- create a new course role with the same name
- enroll a user in a course with the new role
- ensure the enrollments index returns the enrollment for that user
  - /api/v1/courses/<id>/enrollments?role[]=Course+Admin

Change-Id: Icb6e85549b9e39c1ab6f74059a51cbc65a02519b
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/366665
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>