Style button block to not have a text-decoration on hover
ButtonBlock uses the INSTUI Button. When a Button has an href
(which is true with ButtonBlock), INSTUI generates an html <a>.
The canvas common CSS sets the underline color to
var(--ic-link-color-darkened-10); which is not what we sant.
We really want no text-decoratio at all, so I added CSS to
set text-decoration: none;
This solves the wrong color underline by removing the underline,
which was the point of this ticket in the first place.

closes RCX-2796
flag=block_editor

test plan:
  - create a page with Button blocks
  - add a link URL to the button(s)
  - hover over the button
  > expect no underline on the button text

Change-Id: Idba7a4ee6e071113817d346ea51b39dea155e240
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/366686
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Eric Saupe <eric.saupe@instructure.com>
QA-Review: Eric Saupe <eric.saupe@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>