add fixes for rubric form free form comments
this commit closes two tickets. the first ticket is to fix the height of
the rating cells in the traditional rubric view when the criteria is
from an outcome. the other ticket fixes the UI in the rubric form to not
display ratings when the rubric is a free form comment rubric.

closes EVAL-4941
closes EVAL-4942
flag=enhanced_rubrics
flag=enhanced_rubrics_assignments

test plan:
- with FFs on, navigate to the rubric form builder page.
- with criterions created, change the "Type" dropdown to
  "Written Feedback".
- verify that the rating accordion is no longer displayed and instead
  you see text explaining that it is a free form comment.
- create a criterion from an outcome.
- click on the preview button and verify that the rating cell heights
  are the same as the other cells in the traditional view.
- close the preview tray and select "Unscored" from the "Scoring"
  dropdown.
- verify that the threshold value of the the outcome criterion is
  no longer displayed.

Change-Id: I5c5e75bba0be749a977211f881ad313bdcec1f33
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/367124
Reviewed-by: Derek Williams <derek.williams@instructure.com>
QA-Review: Chris Soto <christopher.soto@instructure.com>
Product-Review: Cameron Ray <cameron.ray@instructure.com>
Tested-by: Spencer Olson <solson@instructure.com>