Don't update a translated module name
Fixes FOO-4905
flag=none

When a page is being translated to a different language via
a browser page translator, when a module is edited the name
of the module is translated, and then was saved as translated.

The current code layout doesn't make it easy to capture the
untranslated input field content, so we add a "dirty" boolean
to the edit panel state, and if the input field is not modified
then it is not sent in the API payload when saved, so the back
end will not alter it.

Test plan:
* follow the steps to reproduce in the original ticket
* the "translated name gets saved" bug should no longer occur

Change-Id: Ia09a0cbf2954969c582b2c13b6d5112677e553e6
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/367544
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Adrián Grúber <adrian.gruber@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>