fix: remove dev key domain on edit
Removing the domain from the domain field
on a Developer Key in the manual UI did not
actually removed the domain. It was not present
in the JSON req and the backend did not
updated it, with this fix domain with a null value
will be presented in the api request and
therefore will be updated as well.

fixes INTEROP-8993
flag=none

test plan:
- have an LTI dev key with domain
- edit the dev key in manual config UI
- Clear the Additional Settings->domain field
- Save the dev key and view the dev key again
- Verify Additional Settings->domain is empty

Change-Id: Ieca819e01f5b232509f76d386db728e54cc4366a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/367809
Reviewed-by: Ryan Hawkins <ryan.hawkins@instructure.com>
QA-Review: Ryan Hawkins <ryan.hawkins@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>