check anonymity on submission graphql lookup
We can leak anonymity by allowing a teacher to look up a submission by
assignment id and user id when an assignment is actively anonymous. This
modifies the behaviour to check to see if the user has manage_grades
permission and if the assignment is in an actively anonymous state, if
so, it will not return a submission.

closes EVAL-4978
flag=graphql_honor_anonymous_grading

test plan:
 - Have a course with at least one student
 - Create an assignment with anonymous grading
 - As a teacher, query for a submission for one of the students for that
   assignment.
   query {
     submission(userId: "3", assignmentId: "187") {
       id
       userId
       assignmentId
       anonymousId
     }
   }
 - Make sure there is a no submission returned.
 - Grade all the students assigned to that assignment
 - Run the same query again and make sure the submission is returned

Change-Id: Ifd1b33b7423d3c68f321ba1c00b88d10c81ceab9
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/368673
Reviewed-by: Spencer Olson <solson@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Keith Garner <kgarner@instructure.com>
Product-Review: Keith Garner <kgarner@instructure.com>