Cookieless CSRF protection for /login/oauth2/accept
The SameSite attribute of the current _csrf_cookie is empty,
which defaults to 'Lax' in modern browsers.
When using Canvas with a vanity domain, the OAuth flow in
LTI apps may break because the CSRF cookie is not available.
To solve this issue, we turn off the standard CSRF protection
for this endpoint and introduce a new one which does not use
cookies.

fixes INTEROP-9040
flag=csrf_oauth2_fix

Test Plan:
    Regression test:
    - make sure, the csrf_oauth2_fix FF is off (SiteAdmin)
    - make sure to have a (non-LTI) developer key with
      internal_service=true set
    - launch lti1.3 test tool
    - on the test tool launch page go to the bottom
      open the 'LTI in LTI Test' section
    - choose the 'OAuth Screen' option
    - change the client id to your internal_service DK
    - click Go
    - delete the _csrf_token cookie from your browser
    - click "Authorize"
    - you should get 422 CSRF error
    - do it again, but do not delete the cookie
      it should redirect you to the redirect_url
    With FF on:
    - use the LTI in LTI Test again it should work even
      if you remove the _csrf_token cookie

Change-Id: Ief9d5237facbe83b60dd05d5f57d1de96cc8a761
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/368967
Reviewed-by: David Varga <d.varga@instructure.com>
QA-Review: David Varga <d.varga@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Norbert Renyi <norbert.renyi@instructure.com>