Give only integer values to Intl.DurationFormat
Fixes FOO-5090
flag=none

This was an interesting one...there is nothing in the
MDN documentation for Intl.DurationFormat that requires
all the component properties in the `duration` object
be Integer values; only that they should be numbers.

But at least in the Chrome implementation, passing a
fractional value like 12.95737 for the `seconds` causes
an error to be thrown from deep inside .format(). I
suspect this is actually a bug in the implementation of
Intl.DurationFormat, but in any event it is easy enough
to work around simply by rounding the seconds value.

Test Plan:
* Page views display still looks correct

Change-Id: I553fd104ccb885d06cc292c39f866d476a00622f
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/369177
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>