Fix HTML display issue in submission comments
If the `RCE Lite in SpeedGrader comments` feature flag is enabled, users
create comments using HTML syntax. When the feature flag is turned off or
the input used for editing does not support RCE, the HTML syntax is
exposed to the user.

This commit ensures that teachers can edit their comments in the gradebook
with RCE if the feature flag is enabled. Additionally, if a comment
contains HTML syntax and the feature flag is disabled, HTML tags will be
stripped to provide a clean display.

flags=none

closes EVAL-4638

Test plan:
1. Ensure that the RCE Lite in SpeedGrader Comments feature flag is
enabled.
2. Create a course with an assignment and enroll at least one student.
3. Navigate to SpeedGrader and submit a comment for the student.
4. Open the Gradebook, select the student, and access the sidebar for the
graded assignment. Edit the comment.
5. Verify that the same Rich Content Editor (RCE) used in SpeedGrader is
available for editing the comment.
6. Disable the RCE Lite in SpeedGrader Comments feature flag.
7. Refresh the page and attempt to edit the comment using the previous
(non-RCE) editor. Confirm that HTML syntax is not exposed to the user.

Change-Id: I23e0a0f6673ec47d892a72c97792eded110ac383
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/369986
Reviewed-by: Spencer Olson <solson@instructure.com>
QA-Review: Daniel Kis-Pal <daniel.kispal@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Cameron Ray <cameron.ray@instructure.com>