Continue work on using React Router for accounts/course settings
Refs FOO-4945
Refs FOO-4922
flag=none

Because some of the rewrites of settings tabs have to appear in
both the Account Settings and Course Settings pages, and because
we are trying to use React Router routes to instantiate them,
there was a bit of an awkward situation with the "Alert" tab,
which was sort of worked around in a previous commit.

This cleans that up considerably, undoing the awkwardness in
ui/shared/student-alerts/react and creating a Route for the
course settings page that is patterned after the one for the
account settings page. Now both route components can be used
to instantiate React components on parts of each settings
page DOM tree. We can move stuff over from the JS bundles for
both pages into the route over time, with the eventual goal
being that the JS bundles go away entirely and the component
rendered by React Router just does all the work.

Test plan:
* Exercise all tabs in both /accounts/:id/settings and in
  /courses/:id/settings for some course
* Everything should work exactly the same... no changes to
  functionality were made here, just in how stuff winds up
  on the page

Change-Id: I36f3f609febad0acf2665957ca42e0b48d4fbe17
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/370902
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Adrián Grúber <adrian.gruber@instructure.com>
QA-Review: Adrián Grúber <adrian.gruber@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>