[LX-2278] Weighted Assignment Duration distribute remaining days
When applying weighted assignments durations, the remaining time must
be distributed evenly between the items which types were not weighted.

Example: Start date is March 5 and end date is March 25, so there are
20 calendar days. In weighted assignments tray a weight is set for
assignments: 5 and quizzes: 5, but not for discussions and pages. So when
weightings are applied, the weighted value must be set for items
of type assignmens and quizzes, for discutions and pages the remaining
9 days must be distributed evenly (Start date is ignored in duration
calculation, that's why having 20 calendar days, the total duration
will be 19).

Test plan:
 - Set weighted assignments and apply them:
	* Verify duration is set as the weighting for the elements
 	assigned.
	* For types that were not assigned in weighting tray, the
	remaining duration must be distributed evenly.
	* Las item duration must not be greater than end date.
	* Skipped days and blackout dates are taken into account in
	the calculation.
 - Existing functionallity must remain the same:
	* Select start or end date must calculate the duration of all
	items ignoring weighting assignments previously applied.
	* If individual duration of any item in list must recalculate
	time to complete and change end date.
	* Changing time to complete will calculate the duration of all
	items ignoring weighting assignments previously assigned.

flag=course_pace_weighted_assignments

Change-Id: I22699a3f60f12540e7b9469b708293e4487081ea
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/371882
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Brian Park <bpark@instructure.com>
Reviewed-by: Drake Harper <drake.harper@instructure.com>
QA-Review: Drake Harper <drake.harper@instructure.com>
Product-Review: Drake Harper <drake.harper@instructure.com>