Remove skipped integration test that's no longer needed
Closes FOO-5185
flag=none

I couldn't make this test work when rewriting the Quiz IP
filters feature and skipped it. Thinking about it again, it
is only checking that the popup help tooltip functions, and
that is entirely local to the component itself. Thus there
is not really a need for an integration test of it. There's
an existing component unit test, which will suffice.

Test plan:
* none

Change-Id: I90109a8375d9b99ec1654128e9c3b4564fa7ba9d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/371979
Reviewed-by: Maya Tyner <maya.tyner@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>