Add better screenreader text to "Add Quiz IP Filter" button
Fixes FOO-5197
flag=none

To comply with WebAIM, the button to add a new Quiz IP address
filter on the account settings page needed to be specifically
aria-labeled with screenreader text. The actual text of the
button is just the word "Filter" (adorned with a plus sign to
indicate "add new). That's not very descriptive to someone
relying on a screenreader, and the aria-label text more fully
describes the function of the button.

Test plan:
* Go to /accounts/:id/settings#tab-settings
* Use a screen reader and verify that when the "+ Filter"
  button is selected in the "Quiz IP Address Filters" section
  the screen reader reads out the full text of what the
  button's function is, not just the word "filter" which is
  what is visually displayed

Change-Id: I30dd61cd1e10e40e19438a64645b3b7c50ce4648
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/371989
Reviewed-by: Maya Tyner <maya.tyner@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>