api and model changes for ai_rubrics
The initial approach is to pass a param requesting llm generated
criteria instead of passing the criteria array in the request body. I'm
leaving this API param undocumented for now until we're sure this is the
approach we want to take long term.

We're reusing as much of the existing model creation logic as possible,
and just splitting for whether we parse the user provided criteria data
or generate that data from an LLM

test plan:
- (full test plan coming in UI commit)
- see spec
- you could try manually testing this by crafting a curl request to the
  rubric create API endpoint with the appropriate params set, or wait
  for the UI commit for a full integration test plan.

closes EVAL-5150
flag=ai_rubrics

Change-Id: Iff3d610ddebef49a5ed4b532a2989664cbf93a6f
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/372406
Reviewed-by: Spencer Olson <solson@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Simon Williams <simon@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>