backport jQueryUI #6016 fix
note that we fixed this issue in our patch of jQuery 1.8
in g/3873 but this regressed when we upgraded to 1.9
in g/340322

also note that this change doesn't allow HTML elements in
jQueryUI dialog titles (because, going forward, jQueryUI
doesn't either). g/3873 allowed it by explicitly providing
a jQuery object as the title attribute. I found just one
instance of this in an exhaustive search of the code
(a warning icon), and this patchset removes it.

test plan:
 - see SEC-12024
 - if you edit a course calendar event, check the box to
   use different dates for each section, and leave at least
   one section blank, the warning dialog should not have
   "[Object object]" as its title

flag=none
refs SEC-12024
closes FOO-5260

Change-Id: Ie0401fe0339ee781a055c59694802fc2710eb32b
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/372705
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Adrián Grúber <adrian.gruber@instructure.com>
QA-Review: Jeremy Stanley <jeremy@instructure.com>
Product-Review: Jeremy Stanley <jeremy@instructure.com>