guard against non-existent s3 objects on migrate_attachments
When user files get moved from one user to another, primarily during a
user merge/split, we migrate and copy attachment content from the
source user to the destination user across shards in this case. Most of
the time we end up with duplicates and the conflicts get deleted and
marked with a replacment attachment id. This is all fine and dandy, but
the current attachment is still in a file_state of "available", so we
attempt to migrate the attachment content from the source user to the
destination user, but the attachment content (in some cases) is not
available in S3, so we get the Aws::S3::Errors::NoSuchKey error.

closes FOO-5253
closes CANVAS-PQ6C
closes CANVAS-PPTQ
closes CANVAS-PQ34
flag = none

test plan:
 - set up MRA locally
 - with an s3 file backend,
   - do a consortium home account user merge,
     and ensure the user's files make it to
     their new account
   - verify image files and ensure thumbnails are also
     preserved

Change-Id: Ide9a2b6b4be4e4c745d72db697fcab8658b36c7c
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/373183
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>