Make drag and drop size consistent
before, the drag and drop would be the size of the files
table, which looked strange if there was a large amount of
files. This makes it so that the drag and drop area
is always the same size.

closes RCX-3166
flag=files_a11y_rewrite

test plan:
- go to the files page, have a folder with a lot of files
- drag a file over and ensure that the drag and drop area is at
  the top and is a consistent size, no matter the table size.
- ensure the drag and drop area looks correct when there are
  no files in a folder or only a few files (1-2)

Change-Id: Ib10f8d405631e6c7d13958b452f9861a9a9aa6e4
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/373320
Reviewed-by: Eric Saupe <eric.saupe@instructure.com>
QA-Review: Eric Saupe <eric.saupe@instructure.com>
Product-Review: Kata Virag Keri <kata.keri@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>