Restore "Skip to Content" tool for InstUI sidenav
Fixes FOO-5303
flag=none

The React Route for the InstUI implementation of the Canvas
side navigation was stomping on ALL of the ERB-rendered side
navigation elements, including the "Skip To Content" accessibility
tool. The fix is to render the InstUI side nav in an inner div,
leaving the original "Skip To Content" link to appear and function
as it always has.

Some day we will have to come up with a purely React/InstUI
implementation of "Skip To Content".

Test plan:
* Be sure you have the "New InstUI navbar" feature option turned
  on (this allows the use of the InstUI global navigation at all)
* Visit several Canvas pages (especially courses, users, and
  permissions, which have implemented InstUI nav)
* Start on your browser's address bar and repeatedly tab over
  elements until you enter the page content
* The "Skip to Content" should suddenly appear on top of the
  side navigation, and function as always

Change-Id: Iaff54ff81181822999921b33c394d8efa1c4df9a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/374409
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
Reviewed-by: Adrián Grúber <adrian.gruber@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>