rescue failed instFS deletion
refs EVAL-5238
flags=none

InstFS sometimes does not have record of
the zip files created when downloading submissions
from the assignment page. This causes the user to
not be able to download submissions because it
just fails when instFS can't find the file to
delete if stale. This would cause a SW and make us
manually call destroy_perminantly! on the
attachment to get them unstuck. Now it should
still go on to delete the attachment in canvas to
allow a new submisison.zip to be downloaded.

Test Plan:
- submit to an assignment as a student
- Download submissions once as the teacher
- Submit to the assignment again as a student
to make the download file stale for the teacher
- modify delete_file method within `inst_fs.rb` to
 always throw the deletion error:
  def delete_file(instfs_uuid)
    raise InstFS::DeletionError, "received error"
  end
  (this is to avoid having to setup instFS)
- Try downloading submissions again as the same
user and ensure it is successful

Change-Id: Id53ca7dfc5efad381da904e9f34a0766b005d4f0
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/374613
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
QA-Review: Kai Bjorkman <kbjorkman@instructure.com>
Product-Review: Cameron Ray <cameron.ray@instructure.com>