Migrate CustomData data to jsonb column
CustomData and its API is a robust system to store user-specific
attributes, but the data is currently stored as a serialized hash in
a text field. This is the first commit of a two-part migration to
store data in a jsonb field instead, which will allow more complex
and efficient queries against CustomData attributes.

This first commit creates a new jsonb column in a predeploy. The app
code is updated to read and write to/from both columns. A postdeploy
kicks off a job to copy `data` to `data_json`.

Part 2 (in a later release) will drop the original column.

closes CLX-1241
flag = none

Test plan:
 - Create some CustomData rows using the API (if you don't have any)
   https://canvas.instructure.com/doc/api/users.html
 - Run migrations
 - In a rails console, query your custom data (`CustomData.all`)
 - Expect each row's `data` and `data_json` to match
 - On one object, change `data` and/or `data_json`
 - Expect `data` and `data_json` to be syncronized on save
 - Force `data_json` to be something different by running
   `custom_data.update_column(:data_json, {"something" => "else})`
 - Expect `custom_data.data` to still output the value of `data_json`

Change-Id: Iee10f6683aa4943142111a79f53da43e2b494502
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/375414
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Kevin Qiu <kqiu@instructure.com>
QA-Review: Kevin Qiu <kqiu@instructure.com>
Migration-Review: Cody Cutrer <cody@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>