Resize 'Title' column BulkEditDates
closes EGG-317
flag=none

This shrinks the width of the due date
columns which allows the title column
to be wider.

Note: I needed to change a selenium test selector
in order for this to work. For some reason changing
the width of the due date column caused this
selector to break. I decided to use a data-testid
instead

Test Plan:
- All original tests pass
- Navigate to "Assignments" page
- Select kebab menu at the top right corner
- Select "Edit Assignment Dates"
- Ensure that the title column is wider than before
  - And most assignment titles can fit in it without
    being truncated

Change-Id: I1fc617ff67be0f5b8c7a0a307edf11adc39ce5c1
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/375990
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Martin Yosifov <martin.yosifov@instructure.com>
QA-Review: Martin Yosifov <martin.yosifov@instructure.com>
Product-Review: Sam Garza <sam.garza@instructure.com>