Preselect lti version if LP has no canvas config available
If a tool on the discover page does not have configuration info
available, then when the user clicks Configure, we want to open
the manual tool install wizard and have it preselect the 1.1 or
1.3 radio button.

The 1.1 or 1.3 button will be chosen based on what's in the
product.tool_integration_configurations object.

We want to ignore things inside of that where their
`integration_type` is either "lti_11_legacy_backfill"
or "lti_13_legacy_backfill". If it has other integration_types,
we'll use those to decide if it's considered a 1.1 or 1.3 tool.

fixes INTEROP-9212

flag = lti_registrations_discover_page

test plan:
- Have lp-mock installed
- In lp-mock, edit /api/tools/8a809e85-c29d-4c7f-8ee4-f6c567c2c90a.json
  and delete what's in `canvas_lti_configurations` so that it's just an
  empty array
- In Canvas, have the lti_registrations_discover_page feature flag on
- Go to Admin > Apps
- Click on "LTI 1.3 Test Tool (Global)" (that's the one you just edited)
- Click Configure
- The 1.3 radio button should be selected
- Go back to the Apps - Discover page
- Click on Study.Net
- Click Configure
- The 1.1 radio button should be selected

Change-Id: Ic3ac28e287ffaf81b77deb79c97aee2a898168cb
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/376330
Reviewed-by: Paul Gray <paul.gray@instructure.com>
QA-Review: Paul Gray <paul.gray@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>