Add LTI Asset Processor Settings launch
When an Asset Processor is attached to a
saved assignment, users can now click on
the index menu of the AP and select the
"Modify" menu item. This action opens a
modal, sending an `LtiAssetProcessorSettingsRequest`
to the tool.
If the asset processor was created with a
"window" defined in the content item,
the settings launch will occur in a new tab
instead of the modal.

closes INTEROP-9204
flag=lti_asset_processor

test plan:
- create a new assignment
- attach asset processor
- check if 'Modify' menu item is not present
- save the assignment
- edit the assignment
- check if 'Modify' menu item is present and click it
- LtiAssetProcessorSettingsRequest launch should start
  in a modal
- attach a new asset processor to the assignment
  with this window settings of the content item:
    {"width": 900, "height": 900}
- save the assignment
- check if this new AP's index menu/Modify has an external
  launch icon and when clicked, it launches the tool in a new tab
- attach a new asset processor to the assignment
  with this iframe settings of the content item:
    {"width": 300, "height": 400}
- save the assignment
- check if this new AP's settings launch iframe has the
  specified size

Change-Id: I20f3b5f24ad2e67f70691049619a7e2e48338931
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/376427
Reviewed-by: Ryan Hawkins <ryan.hawkins@instructure.com>
QA-Review: Ryan Hawkins <ryan.hawkins@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>