increase robustness of CourseProgress
with a recent change in CourseProgress, errors started to pop up when
a student was enrolled in a course, and then shortly after a course_index
call was called for that student with course_progress included. this
pattern was specifically observed with Catalog API calls.

this change handles the case when there is a CourseProgress evaluation
requested, but the enrollment is not yet available from the CourseProgress
call to ContextModule. in this case the module will be evaluted as not
completed.

closes DOMSVC-228
flag=none

Test plan:
the issue is not reproducable locally -- passing CI is enough

Change-Id: I804a3e1d6b38077dd97c638869b777862aeffd69
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/376449
Reviewed-by: Drake Harper <drake.harper@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Endre Berki <endre.berki@instructure.com>
Product-Review: Endre Berki <endre.berki@instructure.com>