Add support for OpenDyslexic font
closes LX-2608
flag=use_dyslexic_font

This creates a new brandable_css axis, so the number of configuration
combinations goes from 4 to 6. We do not need the _rtl variants
because the OpenDyslexic font does not incude Hebrew or Arabic
characters.

Because the use_dyslexic_font feature flag is a user flag, it
will take some effort to test this in beta | prod before
turning it on.  The flag currently has `shadow:true` set.
This hides the feature from the UI, even when unlocked.
To test for a user:
1. in site_admin, unlock, but leave disabled
2. use the api to turn the flag on for a single user
   PUT /api/v1/users/15/features/flags/use_dyslexic_font
   sending as the payload
   {"state": "on"}
3. test the user.
   a. do they use the font?
   b. does the rce use the font by default?
4. use the api to turn it off
   PUT {"state": "off"} this time
5. in a subsequent commit, turn `shadow: false`
   Then it will show up on the user profile settings page

OpenDyslexic fonts were downloaded from https://opendyslexic.org/
(actually, when I click on the Download button, I'm taken to
https://antijingoist.itch.io/opendyslexic
where the files are hosted.)
The monospace font only came in an .otf format, so I ran it throught
https://cloudconvert.com/otf-to-woff2 to convert to woff2

The website says the font is being updated continually and
improved based on input from other dyslexic users,
so it might we worthwhile to periodically check back.

As of 2024-07-19, the files are OpenDyslexic 3

test plan:
  - follow the above steps for testing with shadow:true
  - then:
  - in config/feature_flags/learning_foundations_release_flags.yml
    change `shadow: false` for use_dyslexic_font and
    restart your server.
  - unlock the feature at accounts/site_admin/settings
  - log in as a user
  - go to the user's settings (/profile/settings)
  > expect the flag to be visible to the user
  - enable|disabe "Use a dyslixia frendly font
  - refresh the page
  > expect canvas to use the OpenDyslexic font (or not
    dependig o the state of the flag)
  - open a page with the RCE
  > expect the default font w/in the RCE to be OpenDyslexic
  > expect Open Dyslexic and Open Dyslexic Mono to be options
    under Format > Fonts
    (this is true even if the feature is off)

Change-Id: I58f19ae6ef16aa9a05fa486791b652e5b8643997
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/377013
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Reviewed-by: Charley Kline <ckline@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>