lint commits that touch the brandable variables file
This commit provides a [WARN] to the owner / uploader of the commit
that touches the brandable_variables.json file. The intent
is to provide a reminder to the uploader to review the brandable
variables file and follow the instructions via the provided link. Which
should include consideration of the impact of the change on the
brandable variables file and the need to rename the coressponding
migration.

closes FOO-5330
flag = none

Test plan:
- Any commit that touches the brandable variables file should
  receive a [WARN] message to review the instructions in the info

Change-Id: Ifbc24aa2093502e41956052594841083a89c9089
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/377149
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Adrián Grúber <adrian.gruber@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>