Allow schemas to be used for @return values in docs
This implements part of what was in the reverted commit
a089bf398, but in a way that won't cause a problem for deploys.

That previous commit didn't work because it was loading the entire
rails application before generating the docs. I did that so that we
could use objects from lib/schemas in the documentation. However,
this made it crash when the deployable Docker image was being created.

That's because when the deployable Docker image is being created, no
database is present. Therefore, Rails will throw an error if it tries
to load any classes that extend from ActiveRecord::Base. And it runs
`rake doc:api` when creating that image, so we can't include the
entire rails app when running `rake doc:api`.

The solution here is to have the `rake doc:api` task manually
use zeitwerk (rails' autoloading library) and tell it where
to find the lib/schemas directory. This will allow the
documentation generator to use code that's in lib/schemas
(or any other directories we may need in the future) without
loading anything inside of app/models, which will make it
crash.

test plan:
- To mimic how the deploy image is created, do this without postgres
  being up
- `docker compose up -d --no-deps web`
- `docker compose exec web rake doc:api`
- Now bring up postgres so that you can visit Canvas in a browser
- Look at the documentation page for the User endpoints and see
  that the API response for the show endpoint is documented and
  unchanged from how it is in production (aside from User now
  being at the top of the page, whereas UserDisplay was first
  before, which is fine).

Change-Id: I12c04c415a2473c51755823b533723c2a744fe9a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/377154
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Evan Battaglia <ebattaglia@instructure.com>
QA-Review: Evan Battaglia <ebattaglia@instructure.com>
Product-Review: Tucker Mcknight <tmcknight@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>