ensure focus back to +DevKeys button
why: As part of the 5VPAT accessibility audit, it was found that the
focus was not being set back to the +DevKeys button after the modal was
closed. This is important for keyboard navigation and screen reader
users.

closes INTEROP-9274
flag=none

test plan:
- Open the Developer Keys modal by clicking the +DevKeys button.
- Ensure that the focus is returned to the +DevKeys button after the
  modal closes using all the possible ways (clicking outside the modal,
  clicking the top-right X button, clicking cancel button).

Change-Id: I8ac0247c079e8905dbda55eff4a32631c1745bb4
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/377762
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: David Varga <d.varga@instructure.com>
QA-Review: David Varga <d.varga@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>