Match LTI PNS handler url with DK redirect uri
Why?
According to PNS specification:
The domain for the handler must be the same as the
Tool's full domain (as specified in at least one of
the Redirect URLs and/or domain property for the Tool's
registration).

Before this change, the handler domain had to match the
tool domain. With this update, tools can now host their
handlers on different domains as long as those domains are
added to the redirect URIs of the developer key.

fixes INTEROP-9383
flag=none

test plan:
- Regression test:
   - lauch with test tool
   - subscribe to a notice type
- Modify test tool notice_handlers_controller:handler method
  to return "http://localhost/handler"
- subscribe should fail
- add the "http://localhost" to DK / redirect_uris
- subscribe should work

Change-Id: Idb193479c9e6500759f31797d85a611e7670a212
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/379089
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: David Varga <d.varga@instructure.com>
QA-Review: David Varga <d.varga@instructure.com>
Product-Review: Csaba Csuzdi <csaba.csuzdi@instructure.com>