flamegraph service: create html file upon error
flag=none

Adds error handling to the FlamegraphService class. If an error occurs
during the generation of the flamegraph, an HTML file is created with
the error message and saved to the flamegraphs dir in the user's canvas
files.

Test Plan:
- specs pass

Change-Id: I42b5cc48488a4e38f6f4b87153620071556d24f5
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/379233
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Chris Soto <christopher.soto@instructure.com>
QA-Review: Spencer Olson <solson@instructure.com>
Product-Review: Spencer Olson <solson@instructure.com>