remove "Move Contents" menu item option for empty modules
when a module has no items, the "Move Contents..." menu
option is now removed. It reappears when items are moved
back into the module.

fixes LX-2691
flag=module_perf

test plan:
1. create a course and add two modules
2. check that "Move Contents..." does not appear in
the popover menu
3. add an item to one of the modules
4. check that "Move contents..." reappears
5. drag and drop this item into the other module
6. check that "Move Contents..." is visible
in the module with the module item and that it is
not visible in the module with no items
7. test visibility of "Move Contents..." by using
the "Move to" option and "Move Contents..." items in
the popover modals.
8. remove the module item and check that "Move Contents..."
is no longer visible int he popover menu.

Change-Id: I182f41735e839be8f207166c00f44799fdc05ac8
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/379416
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Ed Schiebel <eschiebel@instructure.com>
Product-Review: Zoe Lubitz <zoe.lubitz@instructure.com>