change observer ref to bottom element in Term select
listRef does not run on every change to the elements in a list
and only when the underlying <ul> changes. This caused the intersection
observer to only be set once.

now we set the observer ref to a dummy element at the bottom
of the select options.

fixes FOO-5451
flag=none

test-plan:
- create >100 terms
- visit accounts/<id>/sis_imports
- click full batch checkbox
- click terms Select
- EXPECT the terms Select to lazy-load terms

Change-Id: I4da3aab7e0e3f84c215ff5111ccbac1f963ffe60
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/379781
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Adrián Grúber <adrian.gruber@instructure.com>
QA-Review: Maya Tyner <maya.tyner@instructure.com>
Product-Review: Maya Tyner <maya.tyner@instructure.com>