fix HistoryList pagination
history_controller#index filters the paginated results, and as a result,
it is possible that certain pages return empty arrays, even if there is
a link to the next page. This caused a problem, because in those
cases, the IntersectionObserver could not be attached to any item. This
has been solved by attaching the IntersectionObserver to an empty div at
the end of the list.

fixes FOO-5473
flag=none

Test plan:
 - PREREQUISITES:
   - set up PV4 (or a mock server)
   - Set enable_page_views to pv4 at Site Admin -> Internal Settings
   - set up some PV4 mock data with empty pages in between
 - STEPS:
   - click on History on the Global Navigation
 - CHECK:
   - the pagination should continue on the empty page results

Change-Id: Id661a9e7ac2fa848931a096fb402f1a808465e1c
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/380265
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Adrián Grúber <adrian.gruber@instructure.com>
QA-Review: Imre Foldes <imre.foldes@instructure.com>
Product-Review: Imre Foldes <imre.foldes@instructure.com>