support for windowTarget in some LTI placements
When a tool sets the windowTarget: _blank, the behavior is currently
inconsistent. This has led to confusion for the app developers. In some
cases a new window opens, in other cases it does not.
This change ensures that the windowTarget is respected in the
global_navigation and the user_navigation placements.

closes INTEROP-9157
flag=open_tools_in_new_tab

Test Plan:
We'll be checking the following placements:
  - global_navigation
  - user_navigation

Using the LTI 1.3 test tool...
- Make sure the placements are without the windowTarget setting.
  When launched the tool should open in the same window.

- Make sure the placements have the setting `"windowTarget": "_blank"`.
  When launched the tool should open in a new tab with display type as
  borderless.

- Verify that the documentation for the windowTarget setting is updated
  in doc/api/navigation_tools.md

Change-Id: Idef27a7693bee94b80baea6aced3e656b0ac9cac
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/380376
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>