spec: loading all assets while test runs to fix flakey test
[skip-crystalball]

note: the root cause of the issue was because of
chunk JS files where sometimes it took time to load the file.
the deflater middleware worked in my local. But not on pipeline build.
So now we are loading asset in the beginning when the server starts.
This fixed the issue.

fixes RCX-3504
flag=none

test plan:

Ran Build 7 times with FSC passing 100% (210 times out of 210) in 7 below runs

 Main Build - 136102  & FSC Build - 63382
 Main Build - 136102  & FSC Build - 63383
 Main Build - 136102  & FSC Build - 63384
 Main Build - 136102  & FSC Build - 63385
 Main Build - 136102  & FSC Build - 63386
 Main Build - 136102  & FSC Build - 63387
 Main Build - 136102  & FSC Build - 63388

Change-Id: I61966bb712f4630ad9ec770efd784f8f42308433
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/380901
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: James Butters <jbutters@instructure.com>
QA-Review: Vijay Kannan <vijay.kannan@instructure.com>
Product-Review: Vijay Kannan <vijay.kannan@instructure.com>