allow clear sis stickiness for unchanged user objects
When a user object is unchanged and was previously created via API
or in the UI, we should still allow an explicit clear of SIS stickiness
for those attributes that are "sticky". This is useful for cases where
a user object has been previously marked as sticky, but we want to
remove that stickiness without changing any other attributes of
the user. We are currently already doing this for pseudonyms.

fixes FOO-5474
flag = none

Test plan:
• Create a user in the UI or API
• Create a users.csv with the identical data
• Run the import with the Override UI changes and Clear UI-changed
  state options
• Change the first or last name of the user in the users.csv and run
  the import without overriding UI options
• Verify the name was changed

Change-Id: Ia1a76af85ee44459069aa90f6d07ed9d1a63d0f6
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/380981
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Adrián Grúber <adrian.gruber@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>