Ensure proper loading time of backbone code
Fixes FOO-5509
Reverts g/380868
flag=none

The previous fix for this bug did not address an issue
with how React Router does its "lazy loading" of components.
Empirically (via generous console.log messages), the
createBrowserRouter() function which creates all the lazy
routes does in fact execute all of the import functions in
the `lazy=` props IMMEDIATELY upon execution. Apparently
the "lazy" part is that it does not actually render the
React components themselves until the route is executed.

This finding would seem to create some additional research
work to see if loading the React Router could be optimized
a bit more, or if we are just not making the best use of it
for the way we are invoking it. I don't think we want to
load every single bundle's index.tsx file every time!

In any event, this fixes the immediate problem by not
executing the Backbone.js initialization code until frd
after the corresponding React component has run (which
renders null in this case anyway). That ensures that all
the translations are available when the Backbone and
Handlebars code does execute.

Also caught that the "Edit Assignment" page title was not
being translated, ever, because of the way the original
Backbone and Handlebars code was trying to call I18n.t on
a previously-computed variable (which will never work).

Test plan:
* Switch to a non-English locale
* Edit an assignment (/courses/:id/assignments/:id/edit)
* The ENTIRE page should be in the locale language

Change-Id: Ida9c5fb21f8a07e0ddb72d9a3f09fdc4a98bf633
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/381053
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Adrián Grúber <adrian.gruber@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>