fix classic quiz enhanced rubric outcome creation
this commit fixes the issue where enhanced rubric outcomes dialog was
not opening when creating a new outcome criterion in classic quizzes
assignments page. The issue was causing the UI to break. There is now a
try/catch around opening the dialog, and if it fails, it will display an
error message to the user instead of breaking the UI. The fix ensures
that the needed ENV variable is set from the backend so that the dialog
can open correctly.

closes EVAL-5547
flag=enhanced_rubrics
flag=enhanced_rubrics_assignments

test plan:
- create a new classic quiz assignment
- add or create a rubric
- click on the "Create from Outcome" button
- verify that the outcome dialog opens correctly and that you can select
  an outcome to create a new criterion
- save the rubric and ensure it is saved correctly
- make sure the outcomes dialog also works in for a regular assignment

Change-Id: Iabae3ca32f5e6faf1c28ceadad13a6e01354c7ac
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/381577
Reviewed-by: Derek Williams <derek.williams@instructure.com>
QA-Review: Chris Soto <christopher.soto@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Cameron Ray <cameron.ray@instructure.com>