Ensure proper loading time of backbone code in wiki_page_show
Closes FOO-5534
Reverts g/378776
flag=none

This is the same issue as in FOO-5509, whereby the use of React
Router to load a page which contains no React content but only
Backbone.js code causes the Backbone instance to be instantiated
too early, before the front-end startup code is fully ready for
things to render (in particular the I18n translation pack).

The problem here is that more than one controller loads the bundle,
so until we have a fix for the polymorphic load for "wiki" in
courses_controller, this isn't a good candidate for the blanket
replacement with React Router.

Did manage to get rid of a bit of jQuery, en passant.

Test plan:
* automated tests pass
* be in a non-English locale
* /courses/:id/pages/:pageId still works normally
* Make sure all the buttons on that page are properly translated
* (may be hard to reproduce in dev as this is a race condition)

Change-Id: I4f3e0f62ea1abf40dec4c0cb0e6e013cbf1bcea7
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/382512
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Maya Tyner <maya.tyner@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>