Fix Clearing search results in Exception Modal
Previously the search results were not cleared when the user selected
a context from the search results, even though the search input field
was cleared. This is because the value used to send to the backend was
debounced.

This commit ensures that the state value used to send to the backend
is also cleared when a context is selected from the search results.

fixes INTEROP-9476
flag=lti_registrations_next

Change-Id: Iedd2e0c9d4714410eb00614642faefc5147eec3f
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/382674
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Paul Gray <paul.gray@instructure.com>