Allow "enabled" field to be "true" or "false" string
On the backend, we allow "enabled" to be a string ("true" or
"false"), and will still save it that way. However, our zod schemas
on the frontend are more strict than that, and will error of the
tool configuration is saved with "true" or "false" instead of a
boolean.

Until we have done INTEROP-8921 and made the backend schema more strict,
we need to allow "true", "false", and a boolean to be used for the
"enabled" attribute.

flag = lti_registrations_next

Test plan:
- manually edit the tool configuration of an installed tool and change
  once of the placements to say { enabled: "true" } (with a string
  instead of a boolean
- See that you don't get the "Oops, something went wrong" error when
  viewing that tool in the Account > Apps > Manage page

fixes INTEROP-9402

Change-Id: Ic0085e75fcdff6cf10d0276957bd4035a7dc808e
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/382717
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ryan Hawkins <ryan.hawkins@instructure.com>
QA-Review: Ryan Hawkins <ryan.hawkins@instructure.com>
Product-Review: Tucker Mcknight <tmcknight@instructure.com>