Rewrite ui/features/messages in React/InstUI
Closes FOO-4854
flag=none

This was some VERY old code that used jQuery, jQueryUI,
and Handlebars to implement a SiteAdmin-only feature.
To get rid of the security issues in jQueryUI and the
age of jQuery itself, this was completely rewritten in
React and rendered in InstUI.

I chose to keep the original ERB-rendered tab content
and just "harvest" it out of the DOM to insert into
the InstUI <Tab.Panel> content to save work. This is
not in general an end-user-facing page so strict
adherence to our design guidelines is a little less
important.

There was also a fair amount of breakage in the old
code that I noticed (From addresses never being
provided in a reply, the reply submission returning
BOTH an error and success message in the event of an
API error, etc), and fixed these as well as I could.

Last but not least, added InstUI TopNav and cleaned
up the way this page is loaded by React Router.

Test plan:
* Visit /users/:id/messages on a user id that has had
  at least a couple messages delivered to it
* The display of the messages should still be as
  desired, with the three different views available
  in tabs
* On the "Metadata" tab, the "Reply" function should
  still bring up a dialog and allow a reply to be sent
  via a form submission
* Make sure that /users/:id/messages/:id also still
  works to view one specific message and that the
  top nav "breadcrumbs" for both URLs looks sane

Change-Id: Ic6d8497864e2f1cc0850785d1563e469fdd35a16
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/383783
Reviewed-by: Adrián Grúber <adrian.gruber@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>