Change wording in Delete App configuration modal
We want to be more clear about why an app is disabled.

test plan:
- Just look at it?
- Optionally, with the lti_registrations_next flag *disabled*,
- have a site admin registration installed
- go to the Admin > Apps > Manage page
- click the three-dots menu and choose "Delete App"
- See that the text now says the app is "locked on by Instructure,"
  and that the confirmation button says "Close."

Note:

All inherited registrations are site admin registrations, since
you can't view the Apps table from a subaccount. Therefore, if your
current account is inheriting a registration, we can safely assume it
is a site admin registration.

The `isForcedOn` function that is being used on line 281 only checks
if the registration is inherited, and not which account it's inherited
from, but that is sufficient.

flag = none

fixes INTEROP-9381

Change-Id: Ic158b9d903ad8554f274860749b2227171ceac13
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/383407
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Scott Morgan <scott.morgan@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>