Add a helper class for tool finding methods
We a couple of methods that are used by both ContextToolFinder (for
finding all tools available in an account or course) and ToolFinder
(for finding a specific tool). We thought it was a little weird for
these two classes to reference each other, so those couple of methods
are now in their own class. Methods that are useful to both tool-finding
classes should go here.

test plan:
- N/A

flag = none

Change-Id: Id9304d757bf9060890669a141e99f7e774109832
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/384581
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Tucker Mcknight <tmcknight@instructure.com>
Product-Review: Tucker Mcknight <tmcknight@instructure.com>