Sort accounts/subaccounts alpha in add course modal
Fixes FOO-5542
flag=none

The Add Course modal uses an older Store-style interface
to retrieve and use the list of subaccounts, and this
got missed when we fixed up other lists of subaccounts to
be sorted alphabetically.  This just fixes that.

I didn't mess with the store at all... it rather naïvely
and blindly returns all pages of subaccount records from
the API, which might pose an issue for customers with huge
numbers of subaccounts. This should be fixed, but that's
a much bigger job not withing the scope of this bug. I'll
create another ticket for that work so we don't forget.

Test plan:
* Have accounts and subaccounts
* From the Courses page, click on "+ Course" to bring up
  the modal to add a course
* Clicking on the Subaccount field to pop up the list of
  accounts and subaccounts, and now all should be sorted
  alphabetically within the tree hierarchy

Change-Id: Ib7213ae5765f30b40411a8b5152e6778a3cb07bf
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/384642
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>