Add Asset Report resubmission for group Assignment
In Lti Asset Processor we need to be able to
retrigger and send a submission notice to a tool
even if submission is part of group assignment.
As AP spec we need to send the notice on behalf of the
student who actually did the submission as part
of the group.
As-Is there were no information in the DB about
the real submitter of a group assignment submission.
That is why the real_submitter_id field was added
to the submissions table in this change.

refs: INTEROP-9642
flag=lti_asset_processor
test-plan:
- Create a Group Assignment with Asset Processor
- As a student submit homework
- Verify that submission records are created with
the correct real_submitter_ids
- Send back a PendingManual review for the
submission notice from the (Lti 1.3 test) tool
- Open up SpeedGrader and hit Resubmit Reports
- Make sure notice contains the same
for_user claim as the original one even if the
representative submission (the one that was showed
in SG for the group) is different from the one that
was originally submitted.
- To test submission lookup based on
earlier Asset Records do the same as above
just clear the real_submitter_id after  submissions.
If there was already asset record for that submission
text_entry or file_upload the resubmission should be OK

Change-Id: I889e4b42f5b4539f6318b4e81601244c282997e9
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/386420
Reviewed-by: Evan Battaglia <ebattaglia@instructure.com>
QA-Review: Evan Battaglia <ebattaglia@instructure.com>
Migration-Review: Cody Cutrer <cody@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: David Varga <d.varga@instructure.com>