Match anywhere in subaccount names in NewCourseModal
Fixes FOO-5615
flag=none

SearchableSelect was designed to only match on leading text
when narrowing the options in the list. When looking for a
sub-account from the list, the desired behavior is to match
anywhere. So we added an extra prop to allow that unconstrained
match.

As an aside, do we want this behavior everywhere in the tree
of account_course_user_search? If so we can revisit this and
just remove the extra prop, and change the default behavior.

Test plan:
* Have accounts and subaccounts
* From the Courses page, click on "+ Course" to bring up
  the modal to add a course
* Typing in the "Subaccount" field should match on subaccounts
  in the list anywhere in the names, not just leading text

Change-Id: I260c4a14bf5e1fbebad05c63bdf25fd89b0c0fb3
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/385118
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Adrián Grúber <adrian.gruber@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>