Show only the relevant insight for context view
In speedgrader if we use the context view, the only insight
which should been visible is the relevant entry for the currently
highlighted reply. If there are multiple replies for a student it
still should only show 1 at a time.

closes=VICE-5379
flag=discussion_insights

Test plan:

- Have a discussion with at least two replies from 1 student
- Generate the insight for the replies
- Check in speedgrader if there are multiple insight shows in
the condensed view
- Switch to the context view
- After loading it should be 1 insight visible
- The visible insight should be the pair of the highlighted reply
- Clicking on the Next/Previous button should change the displayed
insight as well and should be in sync with the active reply

Change-Id: I28d955eb84efed133ae8ed5011aeaeffb12a8743
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/385156
Reviewed-by: Theo Sorriaux <theo.sorriaux@instructure.com>
QA-Review: Norbert Zsolt Bakó <norbert.bako@instructure.com>
Product-Review: Tamás Balogh <tamas.balogh@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>