update grading schemes definition of default
being used as a default should reference whether
it's being used as default for a course or an
account - not if it's assigned to an assignment.
Simply being on an assignment classifies it as
'in use' rather than 'default'.

closes EVAL-5589
flag=none

test plan:
- create an account grading scheme
- apply that scheme to a single assignment
- verify that you can still archive that scheme
- now assign that scheme as a course default
- verify that you can no longer archive that scheme

Change-Id: I5b9b46daccc5da31e0db4f862222fe0d59cbcad6
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/385570
Reviewed-by: Daniel Kis-Pal <daniel.kispal@instructure.com>
QA-Review: Derek Williams <derek.williams@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Cameron Ray <cameron.ray@instructure.com>