add `author_viewable_name` field to comment
This is needed because author is not always visible to the user,
and we sometimes need to show an anonymous name. This ensures
we do not return identifying info in the author field and instead
we can use this string to display the name of the author.

closes EVAL-5500
flag=none

Test Plan:
- specs pass

Change-Id: I682405c5dd56584de59c1267d0edcb376f9da96d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/385604
Reviewed-by: Daniel Kis-Pal <daniel.kispal@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Kai Bjorkman <kbjorkman@instructure.com>
Product-Review: Cameron Ray <cameron.ray@instructure.com>