Ensure assignment due date does not appear if no assigned
If an assignment is assigned to everyone and given dates, but then
added to a module with assignees, then everyone is removed as
assignees. This should ensure that the assignment's due date does
not appear in the index page or in the show page.

closes LX-3145
flag=none

test plan:
- create an assignment and assign it to everyone
- add a due date to everyone
- create a module and add the assignment
- assign the module to a section
- notice that everyone is no longer assigned and there are no due dates
- go to the assignment index page, expect to not see a due date
  on the assignment
- act as a student in the assigned section
- go to the assigned assignment, ensure no due date appears.
- ensure due dates do appear if applicable

Change-Id: I7f4bd85f63c3437a52d8ead9e58b76429dbabf4d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/386390
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Ed Schiebel <eschiebel@instructure.com>
Product-Review: Sarah Gerard <sarah.gerard@instructure.com>