generate permissions docs
currently we have a hand-written table of permissions shoehorned
into the docs for the parameters of the Create Role endpoint.
this table is out of date and unwieldy in that location.

remove it and generate a separate documentation page for it
instead, and link to it from the former location of the table.
this can solve both problems.

test plan:
 - generate API docs and confirm:
   - the Permissions page appears in the Basic section
   - links to it from the Roles API work
 - repeat with OUTPUT_FORMAT=markdown

flag=none
refs FOO-5669

Change-Id: Ibee14389088ac91c75d893a20ce31564a8461f77
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/386510
Reviewed-by: August Thornton <august@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Jeremy Perkins <jperkins@instructure.com>
QA-Review: Jeremy Stanley <jeremy@instructure.com>