Validate URL in field
Validates the field using URL.canParse() which is used in other frontend
forms in canvas for consistency. This is not as strict as the validation
on the backend but is consistent with validations found elsewhere in the
frontend of canvas.

fixes LX-3086
flag=modules_page_rewrite

test plan:
- Click "+" to add a new module item
- Select "External URL" from the dropdown
- Enter "not a valid url" in the URL field
- Verify error message "Please enter a valid URL" appears
below the URL field
- Click "Add Item"
- Verify the modal remains open and item is not created

Change-Id: If8c2ecbe44f3f3851fb69f8bb3c6130608e4fe1b
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/386513
Reviewed-by: Drake Harper <drake.harper@instructure.com>
QA-Review: Drake Harper <drake.harper@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Eric Saupe <eric.saupe@instructure.com>