Add Existing registration to DR Install flow
This commit adds the backend capability for an administrator to
"reinstall" an app using the standard Dynamic Registration flow.

It works by including an "existing registration" in the registration
token procurement step, which allows the backend to add that to the JWT
which is sent to the tool. This is tracked so that when the tool
"installs" itself like normally, an `Lti::RegistrationUpdateRequest` is
created (as opposed to a new registration being created).

test plan:
1. Install a new app using Dynamic Registration
2. Request a registration_token, providing that registration's `id` as
   the `existing_registration` parameter.
3. Verify that the registration token contains the
  `existing_registration`
4. Send an installation request to Canvas using that token.
5. Verify that a registration update request is created in the
   database.
6. Verify that the registration update request can be retrieved using
  the uuid via the `lti_registration_update_request_by_uuid` API
  endpoint.

Alternatively, you can check out the WIP UI changes to test this easier.

flag=lti_dr_registrations_update
closes INTEROP-9646

Change-Id: Ifc8d10aa616e2ea9a5cdcc4ab18687cde1141e2c
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/386625
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Paul Gray <paul.gray@instructure.com>