use the actual account when evaluating :account_allows proc
it turned out the issue is more general. Permissions that
use account in the account_allows proc in
permissions_registry.rb could result in an error when
a feature flag or setting is disabled for the root account
but enabled for the sub account. We experienced an
error on the permission table for that vary reason
which resulted in state where the ACCOUNT_PERMISSIONS
was correctly populated with group permissions but
the ACCOUNT_ROLES didn't contained the permission info
based on the role.

as a fix we consider the actual account if the context is an
Account and refer to the root_account otherwise.

refs FOO-5708
flag=none

test plan:
- find a permission which has `account_allows`
attribute (FF or setting based).
- disable the FF or setting on the root account
level and enable it for the sub account
- the permission table should render w/o an
error for the sub account and the permission
itself should be present in the table

Change-Id: Iad5f835560e9ab8ef0428bd3bfb61c24408b9d67
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/386327
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>
QA-Review: Adrián Grúber <adrian.gruber@instructure.com>
Product-Review: Adrián Grúber <adrian.gruber@instructure.com>