account for auto posting and test students
- We were not accounting for test students for the
posted_at field in the section_grade_posted_state.
This would cause SG2 to mark assignments as not
posted when the tst student did not have a grade.

- We also did not have the logic correct for
whether an assignment is posted for a section or not.
We were requiring every submission to have
a posted at regardless of if they were graded or
not.

- SG2 change: We did not allow the post/hide
grade menu to be in a "No grades to Post"
state if the sectoin didn't have any graded
submissions. We also were not accounting for
whether the section filter was on or not.

flag=none
closes EVAL-5885

Test Plan:
- create a manual and auto posted assignment
- ensure the assigned students are in two sections
- open the auto one in SG2
- the eye icon should be shown as posted
- the menu items should be disabled and show "No grades to Post" and
"No grades to Hide"
- grade a student in one section
- filtered by all sections the eye icon should be shown as posted
- the menu should say "All grades posted" and "Hide Grades" with
hide grades being enabled
- filter by the section of the student that's graded and ensure
the eye is the same and the menu options are the same
- filter by the other section and ensure the eye icon should be shown as
posted and the menu items should be disabled with "No grades to Post"
and "No grades to Hide"
- filter by all sections and hide the grades
- ensure the eye icon should be shown as hidden and the menu items
should be "Post Grades" and "All grades hidden" with a hidden pill
- open the manual posted assignment
- ensure the eye icon should be shown as posted
- the menu items should be disabled and show "No grades to Post" and
"No grades to Hide"
- grade a student and ensure the grade is hidden
- ensure the menu items are "Post Grades" and "All grades hidden"
- post grades and ensure the menu items are "Hide Grades" and
"All grades posted"

Change-Id: I61fd5cdf7c50ef1f711d8eebb3745750dc58a75d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/387089
Reviewed-by: Daniel Kis-Pal <daniel.kispal@instructure.com>
QA-Review: Kai Bjorkman <kbjorkman@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Yasi Afsharnive <yasaman.afsharnive@instructure.com>