remove uncalled showFlashError from useModuleItems
if the fetch fails, the error condition is caught further up
the call stack and a flash alert error is shown. The catch
block here is only if the code w/in the try block throws
an error, which is probably never going to happen.  Even so,
in the originl code showFlashError() returns a function that
was never called, so it was useless.

closes LX-3187
flag=module_page_rewrite_student

test plan:
  - in useModuleItems.ts, getModuleItems function, before the return
    insert
    throw new Error("whoops!")
  - load the modules page
  > expect a flash alert  error "Failed loading module items"
  > expect a console error message "Failed to load module items: whoops!"

Change-Id: I3fec3eb4e76587fa9a47871f5434b52750e830e9
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/387353
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Eric Saupe <eric.saupe@instructure.com>
QA-Review: Eric Saupe <eric.saupe@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>