Fix Graphql syntax errors in Frontend
Why?
g/385981 recently introduced graphql codegen
for types in the frontend. The codegen process
scannes the whole codebase looking for `gql` tags
and generate types for the queries and mutations
defined there. As the codegen also takes schema.graphql as
an input now we have type check for all the defined queries.
This commit fixes the existing inconsistencies and invalid queries,
and also updates the schema.graphql to be up to date with the
graphql rails backend code.

This is a great improvement as now we are going to have
type check for our GraphQL queries and mutations in the frontend.

flag=none

test-plan:
- Update the schema with `rake graphql:schema`
- Run `yarn graphql:codegen` to regenerate the types

Change-Id: Id3a0b285d41002844fa600151606531ee78be759
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/387692
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: David Varga <d.varga@instructure.com>
QA-Review: David Varga <d.varga@instructure.com>
Reviewed-by: Csaba Csuzdi <csaba.csuzdi@instructure.com>
Product-Review: Evan Battaglia <ebattaglia@instructure.com>