[LLM documentation] Add .claude file to differentiation tags
Canvas is a huge monolith. We want to have specific instructions
for smaller bite sized pieces of the code for better interactions

This is an example of a .claude file for a specific feature
In Canvas.

flag=none

Test Plan
1. Just review the documentation for any errors

Change-Id: I82b772b696d3e405eeefdeae0b176d2a4b65720a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/387734
Reviewed-by: Caleb Guanzon <cguanzon@instructure.com>
QA-Review: Caleb Guanzon <cguanzon@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Sam Garza <sam.garza@instructure.com>