add support for discrete column polymorphic associations
closes AE-2704

Basically doing the inverse of our old polymorphic extensions -
it defines an association for the "polymorphic" association that's
not backed by a db column. It actually defines the association
though so that it can have a place to store metadata, and so that
`where(context: thing)` will work. It then manually defines
`context` and `context=` methods to hide the association defined
versions, and are backed by the correctly typed individual
association. Finally, it enhances the individually-typed
association setters to be sure to nil out the other types,
maintaining disjunction.

Note that the `options[:polymorphic]` on the association is no
longer whatever the caller happened to provide, but is now the
canonicalized form of a hash with class name keys to column name
values. Some code was adjusted that reaches into that structure.

Finally, several models were adjusted to use this new
infrastructure rather than their disparate implementations that
each had shortcomings.

Change-Id: I1190d1e8ef3fb17605299de83025739508ba4dcb
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/387942
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jacob Burroughs <jburroughs@instructure.com>
Migration-Review: Jacob Burroughs <jburroughs@instructure.com>
QA-Review: Cody Cutrer <cody@instructure.com>
Product-Review: Cody Cutrer <cody@instructure.com>