Associate "Reload this page" explainer to a11y toggles
Fixes FOO-5756
flag=none

If you toggle the "Use High Contrast UI" or "Use a Dyslexia
Friendly Font" in the profile tray, we display a bit of helper
text under the toggle saying that the page has to be reloaded
in order for the a11y change to take effect. But that text is
not associated to the toggle, so a screenreader user would have
to know to go looking for it in order to see that text that
explains what to do next.

Easy enough to fixed via `aria-describedby` which fortunately
the InstUI component correctly passes along from <Checkbox> to
the underlying `<input type="checkbox">` DOM element.

Test plan:
* Using a screenreader, verify that when those toggles are
  flipped, not only does the "helper" text appear but also it
  is read out by the screenreader when the toggle itself is
  navigated to

Change-Id: Iaff1d7df3167d5e54d00da65ff8e10353d162ed3
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/387971
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Reviewed-by: Maya Tyner <maya.tyner@instructure.com>