Fix href="#" scrollTop behavior when top_nav on
When top_navigation_placement FF is on HTML
anchor elements with href="#" in RCE like Pages
won't scroll to the top.
Why?
Because when FF is off the scrollBar is on the
html element otherwise it is on a drawer layout
element, so the built on browser behavior won't
scroll to top of page when clicked.

closes: INTEROP-9521
flag=top_navigation_placement

test-plan:
- Create a page with a link href="#"
- Click the link, the page should scroll to top
- Enable top_navigation_placement FF
- Click the link, the page should scroll to top

Change-Id: I6fdb73f61e010d58ac2656b460a975de8744f9e2
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/388407
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Csaba Csuzdi <csaba.csuzdi@instructure.com>
QA-Review: Csaba Csuzdi <csaba.csuzdi@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>
Reviewed-by: Steven McGee <steve.mcgee@instructure.com>