Limit which resource links are shown in the API
Previously, if someone made an api call to
/courses/course_id/resource_links/resource_link_id,
we were checking if the use had correct permissions for
that course_id, but we were not checking if the
resource_link_id was actually in that course.

This change limits the base_scope of all queries
so that it is only looking at resource links in the
specified course or its assignments.

This was already being done just for the index endpoint;
now it is being done generically for all endpoints.

Also, the index endpoint was preloading the resource
link's context. That is now also being done for all
queries by adding it to `base_scope`.

fixes INTEROP-9568

Change-Id: I8b73ed5c557ffeef83151e62f18fd02f5f5d7364
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/389411
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Paul Gray <paul.gray@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Paul Gray <paul.gray@instructure.com>
Product-Review: Tucker Mcknight <tmcknight@instructure.com>