use abstract assignment as context in :belongs_to_enrolled scope
note: we need to use AbstractAssignment AR
when searching for assignments because the
actual context could be an Assignment or
a SubAssignment. In the pervious version
the code broke with a no resource found
when we used the SubAssignment id in the
`Assignment.find(context_id)` call

refs FOO-5841
flag=none

prerequisite:
- have a course, an enrolled student,
an observer, and a teacher.
- create a graded discussion with
"assign graded checkpoints" setting
- before submitting anything create an
ObserverAlertThreshold from the console
with alert_type: "assignment_grade_low"
and for e.g "threshold: "60".
- have the Canvas Parent app installed
locally

test plan:
- reply to the discussion with the student
- with the teacher grade the discussion
using the "Grade" tab in a way that the
value passes the threshold.
- in the mobile app open the alert of
for the student
- you should not see the error screen
but the actual observer alert

Change-Id: I4e134763f2e6736091689fb07997a5a76b4701bc
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/389488
Reviewed-by: August Thornton <august@instructure.com>
Reviewed-by: Chrystal Langston <chrystal.langston@instructure.com>
QA-Review: Adrián Grúber <adrian.gruber@instructure.com>
Product-Review: Adrián Grúber <adrian.gruber@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>