Add a new insight_entry_id param for SpeedGrader
Currently, when we invoke SpeedGrader, by pressing the
See In SpeedGrader button, we send the entry_id of the
discussion entry in the URL as a query parameter.

However, this is probably bugged, since we sent the
insight entry id instead as entry_id. To be consistent
with the rest of the implementation, we will send the
entry_id instead.

Also, for the automatic focus and select to work, we need
the insight_entry_id to be also sent in the URL.

closes VICE-5512
flag=discussion_insights

test plan:
Note: This is meant to be tested along with
https://github.com/instructure/speedgrader/pull/903

Pre-requisites:
- Have a graded discussion
- Have multiple replies from the same student
- Open the Discussion Insights panel
- Click on the See In SpeedGrader button

Happy path:
- Verify that insight_entry_id parameter is in the URL
- Verify that the correct insight entry is highlighted
- Verify that the correct insight entry is focused

Check if parameter is not in the url:
- Delete `insight_entry_id` param from the URL
- It selects and focuses the first entry

Check if parameter is invalid:
- Set `insight_entry_id` to `invalid-id`
- It selects and focuses the first entry

Check if insight_entry_id is valid but missing:
- Set `insight_entry_id` to `9999`
- It selects and focuses the first entry

Change-Id: I742031928f468b270895de8096eb8e6fda317c6c
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/389652
Reviewed-by: Roland Beres <roland.beres@instructure.com>
QA-Review: Dora Csolakov <dora.csolakov@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Bence Árpási <bence.arpasi@instructure.com>