[LX-3335] Reload page on invitation accept
When a student accepts an invitation to join a
course from the dashboard, we will now refresh the
page to grab the latest information. This will be
revisited in the future to refresh the data without
using location.reload(). There will be more details
for this in the associated ticket.

I also created keys for the Dashboard's queries
so the code is more consistent and  mantainable

Test plan:
1. On a course, invite a student
2. As the student accept the invitation
3. On another course, invite the student again
4. As the student go to the Widget Dashboard and
   you will see an invitation to join the other
   course.
5. Accept the invitation. Notice the page reloads.
6. Notice you can now see the recently accepted
   course in the dropdowns, and any relevant info
   in the dashboard.

flag=widget_dashboard
closes LX-3335

Change-Id: I99266b98169285307ea781a6c8fb61b2e23c3cc3
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/390058
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Sarah Gerard <sarah.gerard@instructure.com>
QA-Review: Sarah Gerard <sarah.gerard@instructure.com>
Product-Review: Jaime Lozano <jaime.lozano@instructure.com>