Open AUP in a separate window if it is a built-in one
Closes FOO-5737
flag=none

If the Terms of Service on an account is set to one of the
"built-in" ones, then instead of pulling the content out of
what's stored in Postgres, just open a new browser window and
point it at the Instructure Acceptable Use Policy page. This
will decouple the actual AUP contents from what's managed by
Canvas itself.

The presence of an external URL is signified by a Setting
with key "external_aup_url_for_XXX" where XXX is the name of
the built-in AUP type (default, default_v2, or internet2)

If custom Terms of Service are set, then just do as before and
open a modal with the TOS contents as retrieved from the Rails
model as set for the account.

Test plan:
* Have an account with a "built_in" terms of service set (this
  is the default so unless you've set custom terms of service
  it will already be set)
* Go into Rails console and
    Setting.set("external_aup_url_for_default_v2",
       "https://www.instructure.com/policies/acceptable-use")
* Restart Rails or send it a SIGHUP
* Go to the "Acceptable Use Policy" on either the default
  login page or from the footer links at the bottom of the
  Canvas Dashboard page
* Instead of a full-screen modal, your browser should open a
  new window with the instructure.com acceptable use page

Change-Id: I64d00a8ba2fb1a30193984b99b05294ed0a7e3e9
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/390282
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
Reviewed-by: Adrián Grúber <adrian.gruber@instructure.com>