spec: Add notification helper to RSpec
This CR is about adding helper methods to keep
the testing logic in one place for our notification.
Currently we are fetching the notification models.
These helpers add a wrapper around it to have a clean syntax sugar.
With these we expect our test coverage might be improved, because
in tests we need to do pretty similar as it is in the
broadcast_policy implementation.

Change-Id: Id9d4f680dbbabc45852a7a64690c84046af676de
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/390694
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Roland Beres <roland.beres@instructure.com>
QA-Review: Roland Beres <roland.beres@instructure.com>
Product-Review: Balázs André <andre.balazs@instructure.com>