Fix Discussion Insights timeout issue
This change request fixes the issue where discussion insights
are timing out because of long processing time when the discussion
has many entries to be processed.

The issue was caused by a query which was not cached and called
multiple times when we checked which entries needs to be
processed.

closes VICE-5597
flag=discussion_insights

Test Plan:

Not sure if this is possible to test in local environment
as it needs a huge discussion with lots of entries.

And even if you have that, we cannot test it because the roundtrip of
the local database is instant, so we cannot reproduce the
timeout issue locally.

I tested the changes by monkey patching the beta environment
and running the methods manually from the rails console.

The fixes were working, but here is a test plan if you want to test it:
- Connect to the beta institute with SSH (details in ticket)
- Get the discussion topic:
- topic = DiscussionTopic.where(id: 273298).first
- Call the needs_processing? method:
- topic.insights.first.needs_processing?
- Notice, that query is slow and takes for ever
- Monkey Patch the code pieces that are in this ticket
- Call the `needs_processing?` again
- Notice that it finishes in a few seconds now


Change-Id: I1621e2fdcf40f471e7c10e310e28d8591b2cd912
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/390714
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Balázs André <andre.balazs@instructure.com>
QA-Review: Theo Sorriaux <theo.sorriaux@instructure.com>
Product-Review: Bence Árpási <bence.arpasi@instructure.com>