Rollup course service
Why?
There are account level outcomes.
We need to be able to update rollup scores if
a user changes the calc method on one of those
outcomes

If we were to use the original service then a job
would be created for EVERY student in the course.
There is a lot of overhead and endpoint calls associated
with that aporach

Here we optimize to be able to orchestrate a series of jobs
that will limit the scope to all students in a course for a
specific outcome.

closes EGG-1936
flag=none

Test Plan
1. Create a course with assignments and at least 1 new quiz
2. Align at least 2 outcomes
3. Have multiple students with outcome mastery scores
4. Run the service for the course and a given outcome
5. Validate that the Outcome service runs and that it
creates rollups for just the given course and outcome id
6. validate rollup is accurate.

  # Run service for outcome1 ONLY
  service = Outcomes::CourseOutcomeRollupCalculationService.new(
    course_id: course.id,
    outcome_id: outcome1.id
  )
  result = service.call

  # Or async via delayed job
  Outcomes::CourseOutcomeRollupCalculationService.calculate_for_course_outcome(
    course_id: course.id,
    outcome_id: outcome1.id
  )

Change-Id: I1efad63be4964ace0f5be9ae5b81544a88df49f1
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/390653
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Balázs Buri <balazs.buri@instructure.com>
QA-Review: Balázs Buri <balazs.buri@instructure.com>
Reviewed-by: Mark Starkman <mark.starkman@instructure.com>
Product-Review: Kyle Rosenbaum <krosenbaum@instructure.com>