Improve accessibility for NEW pills on help links
Fixes FOO-5798
flag=none

We should associate the pill text with the link itself so that
when using a screenreader, it will read out the contents of
the help link and then immediately the contents of the Pill
(for example "Ask the Community, NEW, button").

This commit associates the link with the pill using an aria-
describedby attribute.

Test plan:
* Have a custom help link that's designated as "New"
* Open the Help menu in the sidenav
* With a screenreader enabled, navigate the Help Links in
  the tray
* When the "new" help link is focussed, the word "New" should
  be read out by the screenreader after the text contents of
  the help link

Change-Id: I38f397338e3a0395bd3f033e1a0cdd03a7319418
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/391126
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Adrián Grúber <adrian.gruber@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>