Support overriding MFE urls at runtime
Creates an endpoint which allows engineers to specify the URLs to use
for microfrontend apps in Canvas, overriding the ones set in
DynamicSettings for their session. When an override is active, an
indication is shown in the Canvas UI. Currently, only Canvas Career
MFEs consume the override, but this is extensible for other Canvas
MFEs.

This is only supported when the
`allow_microfrontend_release_tag_override` setting is set, i.e., so it
can be enabled only in development environments. It also restricts the
asset host to assets.instructure.com.

refs CLX-3024
flag = none

Test plan:
 - Go to /api/v1/microfrontends/release_tag_override
 - Expect a 404 (since the setting is not enabled)
 - `Setting.set("allow_microfrontend_release_tag_override", "true")` and
   restart rails
 - Go to /api/v1/microfrontends/release_tag_override?override
   [canvas_career_learner]=https://assets.instructure.com/test
 - Expect to be redirected to /
 - Expect to see the indicator in the bottom right corner
 - Open Canvas Career and expect an error about not being able to load
   assets from that /test URL
 - Click "Clear Overrides"
 - Expect to be redirected back to the same page and for the MFE
   overrides indicator to be gone

Change-Id: I329f5883f9c701affcf64f99ddbd5d17e2f833ed
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/391218
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>