Update student rollup service to follow cleaner pattern
refs EGG-1936
flag=none

Why?
When implementing the course version of this service we cleaned
up the methods to make it easier to read.
This change brings those updates to the student level service so
that they are consistent and easier to work with.

Test Plan
1. in console run a course and a student
level process.
2. Verify rollups are created

Change-Id: I8cbcfb0647303cce82407baf87e363ae80b4b58b
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/390875
Reviewed-by: Mark Starkman <mark.starkman@instructure.com>
QA-Review: Mark Starkman <mark.starkman@instructure.com>
Product-Review: Kyle Rosenbaum <krosenbaum@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>