rubocop: downgrade Rails/StrongParametersExpect
it unexpectedly "enabled" itself with the Rails 8.0 upgrade, the
autocorrect isn't trustworthy, there are a lot of extant violations,
and it would take too long to fix them when it's causing issues for
people. so just downgrade it and disable autocorrect so we can
get on with our lives. perhaps we'll revisit later when we have
time to understand what the cop is wanting us to do, and how to
do it safely

Change-Id: Ibe72b403dcf57b7f26e6f06b0ed05f4da2f6c24e
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/391795
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jake Oeding <jake.oeding@instructure.com>
QA-Review: Cody Cutrer <cody@instructure.com>
Product-Review: Cody Cutrer <cody@instructure.com>