Fix issue in data fixup where uses with no enrollments were being affected
There was an error in the logic when determining if a user had no enrollments.
Because it was iterating through all shards and then only considering a user to have
zero enrollments if they had no enrollments on any of those shards, we would often
be looking for enrollments on a shard where we didn't expect a user to have enrollments.
That shard was simply added because a user in the same fixup batch had enrollments on that
shard, but we didn't expect *every* user in that batch to possibly have enrollments on
every shard. So they wouldn't be a "no enrollments user" on that shard because they weren't
a user on that shard at all.

This new version checks which users have no enrollments on a shard, then only considers the users
that we thought *might* have enrollments on that shard. Then only keeps those users as candidates
for having zero enrollments in total.

Basically, on any given shard,
"users_with_enrollments_on_this_shard = users_who_could_have_enrollments - users_with_enrollments."
Then remove those users from the list of candidates for having no enrollments.

flag = none

refs INTEROP-9785

Change-Id: I6ea117c25c7c7e9cbe2905fec89b24bc196ce95f
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/392748
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Tucker Mcknight <tmcknight@instructure.com>