Fixup to make registrations' created_at date match their dev keys
When viewing an app registration on the Apps page (with the
lti_registrations_next flag either on or off), the "Installed On"
field shows the created_at date for the registration record in the database.

If the lti_registrations_next flag is off, this is shown in the table on the
Manage tab. If the flag is on, it's shown in the info area when you click the
app for more details.

These dates are incorrect if the registration was created by one of our backfill
jobs that created registrations from existing developer keys. Those dates all
show the date that the backfill happened, since that's when the registration
records were created. The date that the developer key was created should be
considered the real installation date.

To fix that, this finds all registrations whose created_at dates are more than
one hour different from their dev key's created_at date, and changes the
registration to match the date on the dev key.

fixes INTEROP-9528

flag = none

test plan:
- Manually edit the created_at date of one or more registrations (e.g.
  using .update_attribute) to give them a created_at date that is later
  than the created_at date for their associated developer keys.
- Run migrations.
- See that the registrations whose dates you changed now have a created_at
  date that matches the date of their developer keys.
- Note: this should only affect registrations whose dates are more than one
  hour different from their developer keys. This is to minimize the number
  of registrations we are affecting. We really just want them to be within
  the same day so that they "Installed On" column shows the correct date --
  the exact time isn't crucial.

Change-Id: I61e068956634835cc9b9ffc8aeb8260fab6ac53a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/393495
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ryan Hawkins <ryan.hawkins@instructure.com>
QA-Review: Ryan Hawkins <ryan.hawkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
Migration-Review: Weston Dransfield <wdransfield@instructure.com>
Migration-Review: Cody Cutrer <cody@instructure.com>
Product-Review: Xander Moffatt <xmoffatt@instructure.com>