Ensure tool's workflow_state matches its account binding
Before this change, when calling the CreateRegistrationService,
the tool doesn't always get its workflow_state updated to match
the account binding that is created right before the tool.

This was because we usually rely on a delayed job to make the tools
associated for a developer key match their account bindings. This job
is triggered when an account binding is updated. That wasn't happening
when using the CreateRegistrationService to create a tool and account
binding simultaneously.

The fix here is to manually set the tool's availablity to match the
account binding's availability, if the caller specified an availability
when calling the CreateRegistrationService.

Additionally, we need to start showing disabled ContextExternalTools
on the Availability and Exceptions tab, so that the newly-created,
disabled deployment still shows up on that page. (Previously, they
were showing up because of the bug that left them enabled when
they should have been disabled.)

test plan:
- With the lti_registrations_next flag on,
- use the old developer keys page to create an LTI key. (E.g. with the
  lti 1.3 test tool.)
- Do not turn on this new LTI dev key.
- Go to the new apps page and edit the default context control for this new tool,
  making it available.
- See that the tool still doesn't show up anywhere until you turn on the developer
  key.

flag = lti_registrations_next

fixes INTEROP-9864

Change-Id: Icfd6a6ee2cc86f9229b3c7fd338af7e7d2b84d16
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/393603
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Xander Moffatt <xmoffatt@instructure.com>