Reject "unreasonable" dates in assignments
Fixes FOO-5874
flag=none

Long ago we added sanity checks for the year in the old
jQuery datepicker to prevent accidental selection of years
that are obvious nonsense (say, before 1980)... however as
we continue to deprecate the jQuery picker in favor of
the InstUI component we have to be sure that code that
uses it also contains these checks. This at least adds the
checks into Assignment creation/editing, but we'll need to
keep an eye out on other places in the code where this
same check will need to happen.

Test plan:
* Have a course with an assignment, or make a new assignment
* In the Due Date, Available from, and Until selections for
  those dates and times, everything should work as before,
  including "other" validations like the due date being
  before the Available date
* In addition, if you pick a date before 1980 in any of
  those pickers, it should be rejected with an error

Change-Id: Ife08beacc16b11455375371300e8b93daeaf1341
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/393717
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Adrián Grúber <adrian.gruber@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>