further refine oauth2 timeouts
Faraday is catching the Timeout::Error and wrapping it in a Faraday::TimeoutError,
which does not inherit from Timeout::Error.

Also, Canvas::TimeoutCutoff is only ever thrown from outside the block, not inside
it, so we need to catch it outside the block.

refs AE-3113

Change-Id: I19fbbeeec13df7f417d0d8799a10a39e9fb34af1
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/394042
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jacob Burroughs <jburroughs@instructure.com>
QA-Review: Cody Cutrer <cody@instructure.com>
Product-Review: Cody Cutrer <cody@instructure.com>