Prevent editing review settings after PR submissions
This is PS 1 of 3 for EGG-1859.  This PS adds a
validation check for the Reviews required and Points
per Peer Review review settings cannot be changed
after students have submitted peer reviews. This
change includes a validation check on the respected
models for points_possible (PeerReviewSubAssignment)
and peer_review_count (Assignment) that checks for
changes in the values and if there are peer review
submissions (AssessmentReviews) completed then it
throws an error.  The last piece that is added is
a boolean attribute on the assignment called
has_peer_review_submission to indicate if there
are submissions or not.  This will be used in PS2.

refs EGG-1859

flag=peer_review_allocation_and_grading

Test plan:
- Tests are sufficient and jenkins passes
- Enable peer_review_allocation_and_grading feature
  flag
- Create assignment with peer reviews enabled
- Do not submit any peer reviews
- Verify has_peer_review_submissions is returned in API
  - Either by using the browser or POSTMAN perform a GET
    request using the below:
    http://[canvasdomain]/api/v1/courses/[course id]/
    assignments/[assignment with peer review enabled]
  - Verify the value is false to reflect that 0 peer
    reviews have been submitted
- Allocate peer reviews to a student
- Log in as the student and complete the peer review
- As an instructure check the API again and verify
  the has_peer_review_submissions is true
- Validate error is returned when trying to update
  peer_review_count via REST
  - Using postman, create the below request:
  - method: PUT
  - url:
http://[canvas domain]/api/v1/courses/{course_id}/
assignments/{assignment_id}
  - body:
  {
    "assignment": {
      "peer_review_count": 2
    }
  }
  - headers:
    Authorization: Bearer YOUR_ACCESS_TOKEN
- Validate error is returned when trying to update
  points_possible via REST
  - Using postman, create the below request:
  - method: PUT
  - url:
http://[canvas domain]/api/v1/courses/{course_id}/
assignments/{assignment_id}
  - body:
    {
      "assignment": {
        "peer_reviews": true,
        "peer_review": {
          "points_possible": 75
        }
      }
    }
  - headers: Authorization: Bearer YOUR_ACCESS_TOKEN

Change-Id: I1bc4350b33e8565867fe8cec0b37abb7f801dd6d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/394632
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jason Gillett <jason.gillett@instructure.com>
QA-Review: Jason Gillett <jason.gillett@instructure.com>
Product-Review: Jason Gillett <jason.gillett@instructure.com>