Add notice that theme updates take time to roll out
Closes FOO-5450
flag=none

When making a change to a theme in the theme editor, some
customers mistakenly believe that the theme and branding
changes will take place immediately after they click on
the "Apply" button, but actually jobs and caching make
it so that the changes take some time to appear. This
commit adds language to the Apply Changes confirmation
to set that expectation for users.

I also changed the old `window.confirm()` into an actual
InstUI modal dialog to keep with current design language.
And also did a bit of general cleanup of this file, and
converted it to a React function component.

Test plan:
* Go into the theme editor on a theme
* Change something, preview and save
* Click the "Apply theme" button and note the changes
  in the confirmation dialog and new language
* After applying a theme change, also notice the flash
  notice on the main theme page, again as a reminder that
  the changes won't be immediately visible

Change-Id: I689bb0cae3d1e182770d142c8cc6cd3c94e8520a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/394368
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Adrián Grúber <adrian.gruber@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
QA-Review: Maya Tyner <maya.tyner@instructure.com>