fix: change accessibility table default sort to issues descending
closes RCX-4194

This ticket updates the default sorting parameters for the resource
scan result, so that the items with the most issues get to
the top table. The change now only allows 2 states for sorting.

However there is also an AC, which states that, we still need to
support user overrides. As in, if someone has touched the table sorting
at least once, that has to be persisted.

During the page load you can see that the table is initially
sorted by issues descending, but if someone ever touched the
sorting controls, it will then go to that last set "override" state.

flag=a11y_checker

test plan:
- open accessibility page
- table should sort by default to descending issue count

Change-Id: I8594c304908472992b0410beea189fa146db983f
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/394533
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Kadir Saglam <kadir.saglam@instructure.com>
QA-Review: Kadir Saglam <kadir.saglam@instructure.com>
Product-Review: Zsolt Mester <zsolt.mester@instructure.com>