fix asset reports not showing for checkpointed discussions
Submissions for the main assignment of a checkpointed discussion do not
have a submission type until they are considered fully submitted, that
is, until the student has both 1) posted a top-level entry and 2) made
the required number of "additional replies". However, if the student has
created any entries, these submissions can still have asset reports.

We use the submission_type of an submission to help indicate the
possibility of having asset reports (some submission types, such as
online_url, can't have reports). This commit adds the concept of an
"inferred submission type" which is normally just the submission_type,
but is set to be discussion_topic if the submission_type is null but
there are or could be reports.

On the student grades pages, we were already fetching the reports (since
we fetch all reports for a course), so I just needed to override the
submission_type if there are discussions-based asset reports. In
speedgrader, we weren't even requesting reports if the discussion type
is null. Unfortunately, there's no easy way to tell if a student has any
entries, so I had to bypass the optimization -- if the assignment is a
checkpointed discussion, we need to check to see if there are reports.
This ensures we will show any reports that are shown on the student
grades pages.

Test plan:
- have a graded discussion with checkpoints (set additional replies to
  something >= 3) in a course with at least two students.
  alternatively, you can have just less students and set up the test
  cases with multiple assignments.
- have one student post one top-level entry and make an asset report for
  that entry. Make sure it is in a "Processed" state.
- have another student reply to the other student's post, but not make
  any top-level posts. make an asset report for that entry.
- double check in the database that both students' submissions'
  submission_type is null:
    DiscussionTopic.last.assignment.submissions.pluck(:submission_type)
- check that both reports show in speedgrader. Note that for the second
  student, in speedgrader it will say "the student does not have a
  submission", because they haven't satisfied any of the two
  checkpoints. (it still shows the entries on the student submission
  page...)
- check that both reports show in old student grades page, new student
  grades page, and old student submission page
  (/courses/1/assignments/2/submissions/3). Make sure the Processed
  submission shows for the student in all three places. (Old student
  grades page can be easily distinguished from new in that old lists the
  subassignments in an expandable div below the discussion.)
- make a new, non-checkpointed discussion, with an asset processor. Do
  not make any student discussion entries.
- view that discussion in speedgrader. Open the dev tools to the network
  tab and reload the page. Make sure we do _not_ make a graphql request
  to get the asset reports (since an unsubmitted submission for a
  non-checkpointed discussion cannot have reports).
- view the student submissions page for that discussion and make sure
  there that we also do not try to fetch the asset reports from graphql.
- check that a checkpointed discussion without any asset processors (or
  reports) doesn't show anything about "Document processors" on speed
  grader, old and new student grades page, and student submission page

closes INTEROP-9995
flag=lti_asset_processor_discussions

Change-Id: I4fbb6461e14362f80813a8283c1a59711922ecd6
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/394570
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Csaba Csuzdi <csaba.csuzdi@instructure.com>
QA-Review: Csaba Csuzdi <csaba.csuzdi@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>