Implement Grouping in PeerReviewSelector
There were some design decisions I had to make and
might need to be tweaked in the future. But I matched
the ticket and figma as closely as possible.

First, the ordering and name of peer reviews in the selector.
- All just follow "Peer Review x of y" where y is the total
count of available peer reviews. The order is based on creation.
- This can lead to some unintuitive ordering, but there are
complications if you order based on other states due to the
ambiguous name of the peer review.

Second: What happens with assessmentRequests that aren't ready.
- Neither the figma nor the tickets talk about how to handle
this state. I opted to hide them from the selector since
there are no expected actions for a student to take based on
a peer review that is not available.
- I could see this causing some confusion if a student is
assigned 3 peer reviews, when they look at the peer review
page they only see 2 available, and so they think they have
completed their peer reviews when in reality they are waiting
on another student to make a submission.

closes EGG-2163
flag=peer_review_allocation

Test Plan
1. Have a peer review assignment with at least 4 students
2. Assign 3 students to another student to peer review
3. Acting as the student that needs to peer review, go to
the peer review page:
http://canvas-web.inseng.test/courses/15/assignments/124/peer_reviews
4. Verify that it says "No peer reviews available"
5. Act as 2 of the other students and make a submission
6. Go back to the same page as in step 3
7. Verify that 2 options appear as ready to review
8. Complete the peer review of one student
9. Navigate back to the page in step 3
10. Verify that the peer review selector displays 2 peer
review options and groups them: one in ready to review and
one in completed reviews.

Change-Id: I3826a2bb0b079f29bdbcb31139b020c84ce63142
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/394745
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Angela Gomba <angela.gomba@instructure.com>
QA-Review: Angela Gomba <angela.gomba@instructure.com>
Product-Review: Angela Gomba <angela.gomba@instructure.com>