Make footer sticky on SuccessView and remove empty scrolling
On our SuccessView, the footer was at the end of the View, and we
could only see it if we scrolled down.

The container was scrollable, even though the content did not
require scrolling, because there was just empty space.

This fix makes the footer sticky, so it works correctly on
smaller screens and also gets rid of the scrolling bug.

closes RCX-4479
flag=a11y_checker_eap

Test Plan:

- Go To the Accessibility page
- Open an issue and fix it
- The SuccessView with the balloons should come up
- Verify that it does not have any scrollbars
- Verify that footer can be seen at the bottom
- View the component on smaller screen
- The view should be responsive and footer should be sticky

Change-Id: Id3fd0efb280f1d9d900a3cd179d4c2509a279823
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/395292
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Akos Horvath <akos.horvath@instructure.com>
QA-Review: Akos Horvath <akos.horvath@instructure.com>
Product-Review: Norbert Zsolt Bakó <norbert.bako@instructure.com>