Create PeerReviewWidget UI
closes EGG-2181
flag=peer_review_allocation_and_grading

Note: In the designs there is a "Student View" button that was
left out in this PS. This is because this button was not
discussed or scoped for. Will follow up with prod and design
about it in the future.

Test Plan:
- Create an assignment with peer review enabled
- View the assignment as a teacher and observe the Peer Review
  Widget at the bottom
- Widget should include Peer Review text, View Configuration
  button, and Allocate Peer Reviews button
- Disable peer review on the assignment
- View the widget is not shown on the assignment page

Change-Id: Ibc96f9775576c31ddd65af052446c6b276aa5635
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/395640
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jason Gillett <jason.gillett@instructure.com>
QA-Review: Jason Gillett <jason.gillett@instructure.com>
Product-Review: Angela Gomba <angela.gomba@instructure.com>