block role name collision on linked accounts
The current validations ensure that role names will be unique on an
account level. However, when the roles are coming from multiple accounts
through the accounts being linked, it is possible that a role with the
same name will be created, which causes confusion for some customers in
consortiums. On the front-end, there is already a validation for this,
so it makes sense to apply this constraint on API calls as well.

refs FOO-5964
flag = none

Test plan:
  - Create a custom role on a root account
  - Try to create a custom role with the same name on a child account
  through the API
  - CHECK: you should receive a 400 response

Change-Id: Ib1275d66ee4e87f6ccd4b6e91c741272effa6b48
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/396332
Reviewed-by: Adrián Grúber <adrian.gruber@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Imre Foldes <imre.foldes@instructure.com>
QA-Review: Imre Foldes <imre.foldes@instructure.com>