Remove double scrollbar from Dynamic Reg Wizard
This commit removes the double scrollbar issue in the Dynamic
Registration Wizard. The problem was caused by the extra div that was
added to the bottom of the RegistrationModalBody component. This extra
spacing is still needed in other contexts, so we conditionally render it
based on a new, optional prop, `bottomSpacing`, which defaults to true.

test plan:

1. Install a tool using the Dynamic Registration Wizard and verify that
no double scrollbar appears while the wizard is launching the tool.

closes INTEROP-10115
flag=none

Change-Id: Ia7cae96bf05cbb05078ebb58ad113af50f82d540
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/396439
Reviewed-by: Ryan Hawkins <ryan.hawkins@instructure.com>
QA-Review: Tucker Mcknight <tmcknight@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Paul Gray <paul.gray@instructure.com>