Fix decorative image for Alt Text as Filename rule
Currently, we enable nil values for the ImgAltRule and
ImgAltLengthRule rules. However, there is an additional
rule for the filename alt text called ImgAltFileNameRule.

In this case we have the decorative image checkbox on the form,
but when ticking it and trying to save, it returns an error,
because it doesn't accept the null value we are trying to send.

Expected behavior is to be able to set the alt text to null to
signal that this picture is decorative only.

closes RCX-4515
flag=a11y_checker_eap

Test Plan:

- Upload a new image to a wiki page
- This will set the default alt text to the uploaded filename
- Check the issue in the Accessibility page
- Tick the Decorative Image checkbox
- Click on the Apply button
- Click on the Save & Next button
- The issue should be saved and resolved successfully

Change-Id: Ie438e8e6a0fbe5df945f46b23372e2fbf2e28b3f
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/396516
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Akos Horvath <akos.horvath@instructure.com>
QA-Review: Akos Horvath <akos.horvath@instructure.com>
Product-Review: Norbert Zsolt Bakó <norbert.bako@instructure.com>