Fix image display on alt text issues
Currently, in some cases when we show the image for fixing
the alt text issue, it is not properly displayed. There are
scrollbars and also the alignment is also off.

This change fixes the display of images for the alt text issues.
It makes sure that the image is contained in the container and
it is aligned center if the aspect ratio is different from the
container.

closes RCX-4516
flag=a11y_checker_eap

Test Plan:

- Check all 3 alt text issues in the wizard
- In every case the images should display correctly
- No scrollbars and images are centered
- Horizontally and Vertically

Change-Id: If49479e0d6208be75fdbb377d791f9bf8f196d8b
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/396517
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Katalin Kiss <katalin.kiss@instructure.com>
QA-Review: Katalin Kiss <katalin.kiss@instructure.com>
Product-Review: Norbert Zsolt Bakó <norbert.bako@instructure.com>