Make the Tool Configuration Edit button respond to keyboard input
Previously, nothing would happen if you used the keyboard to focus on
the Edit button and then tried to press the button with either spacebar
or enter.

This is somehow caused by the fact that it's an InstUI Button using
"as RouterLink," which isn't something we've done elsewhere.
(Other "as RouterLink" elements are all Link elements themselves.)

The fix here is to make the Edit button be a normal button that calls
react router's navigate() function when clicked. Keyboard users pressing
Space or Enter also trigger the onclick event on a button, so the click
event will capture those keypresses as well.

This does mean that you can no longer right click the Edit button
to open the edit screen in a new browser tab. After discussing it in Slack,
we decided that we were OK with that.

flag = lti_registrations_next

fixes INTEROP-9774

test plan:
- With the lti_registrations_next flag enabled,
- Go to Account > Apps > Manage
- Click on one of your apps
- Go to the configuration tab
- Press the tab key until you have highlighted the Edit button
- Press the spacebar or enter key. It should show you the edit screen,
  as though you had clicked the edit button.

Change-Id: Ideb83597729d81b571f83fd874a4df5771d6946b
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/396903
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ryan Hawkins <ryan.hawkins@instructure.com>
QA-Review: Ryan Hawkins <ryan.hawkins@instructure.com>
Product-Review: Tucker Mcknight <tmcknight@instructure.com>