ignore previous search if search is not in list of options
smart search will reset the search value to the previous
value if the selectedOptionId is still set, but the list
of recent searches is empty. this only happens when typing
the new search query quickly, so it's not consistently
reproducable.

the solution is to leave the input value as-is, instead of setting
to a default "" value in this scenario

also as a little cleanup: hide the recent search dropdown if there
are no recent searches

fixes FOO-5976
flag=smart_search

test-plan:
- pre-req: enable SS
- visit SS with a prefilled query (ex. "q=test")
- do exactly the following and quickly:
	- clear the search field by selecting the text + backspace
	- type in a new query (ex. "quiz")
	- click "Search"
- this can take multiple attempts

Change-Id: I403c134405ebb7ffefa4b8d385395cd63b23352f
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/397022
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>
QA-Review: Maya Tyner <maya.tyner@instructure.com>
Product-Review: Maya Tyner <maya.tyner@instructure.com>